+++ R/license.R
@@ -4,19 +4,28 @@
 #' Adds the necessary infrastructure to declare your package as licensed
 #' with one of these popular open source licenses:
 #'
-#' * [CC0](https://creativecommons.org/publicdomain/zero/1.0/): dedicated
-#'   to public domain. Appropriate for data packages.
+#' Permissive:
 #' * [MIT](https://choosealicense.com/licenses/mit/): simple and permissive.
-#' * [Apache 2.0](https://choosealicense.com/licenses/apache-2.0/):
+#' * [Apache 2.0](https://choosealicense.com/licenses/apache-2.0/): MIT +
 #'   provides patent protection.
+#'
+#' Copyleft:
+#' * [GPL v2](https://choosealicense.com/licenses/gpl-2.0/): requires sharing
+#'   of improvements.
 #' * [GPL v3](https://choosealicense.com/licenses/gpl-3.0/): requires sharing
 #'   of improvements.
 #' * [AGPL v3](https://choosealicense.com/licenses/agpl-3.0/): requires sharing
 #'   of improvements.
+#' * [LGPL v2.1](https://choosealicense.com/licenses/lgpl-2.1/): requires sharing
+#'   of improvements.
 #' * [LGPL v3](https://choosealicense.com/licenses/lgpl-3.0/): requires sharing
 #'   of improvements.
-#' * [CCBY 4.0](https://creativecommons.org/licenses/by/4.0/): Free to share and
-#'    adapt, must give appropriate credit. Appropriate for data packages.
+#'
+#' Creative commons licenses appropriate for data packages:
+#' * [CC0](https://creativecommons.org/publicdomain/zero/1.0/): dedicated
+#'   to public domain.
+#' * [CC-BY](https://creativecommons.org/licenses/by/4.0/): Free to share and
+#'    adapt, must give appropriate credit.
 #'
 #' See <https://choosealicense.com> for more details and other options.
 #'
@@ -29,6 +38,13 @@
 #' @param copyright_holder Name of the copyright holder or holders. This
 #'   defaults to "{package name} authors"; you should only change this if you
 #'   use a CLA to assign copyright to a single entity.
+#' @param version License version. This defaults to version 2 for the GPL
+#'   (for compatibility with R itself), but the latest version for all other
+#'   licenses.
+#' @param include_future If `TRUE`, will license your package under the current
+#'   and any potential future versions of the license. This is generally
+#'   considered to be good practice because it means your package will
+#'   automatically include "bug" fixes in licenses.
 #' @seealso For more details, refer to the the
 #'   [license chapter](https://r-pkgs.org/license.html) in _R Packages_.
 #' @aliases NULL
@@ -52,38 +68,49 @@ use_mit_license <- function(copyright_holder = NULL) {
 
 #' @rdname licenses
 #' @export
-use_gpl3_license <- function() {
+use_gpl_license <- function(version = 2, include_future = TRUE) {
+  version <- check_license_version(version, 2:3)
+
   if (is_package()) {
-    use_description_field("License", "GPL-3", overwrite = TRUE)
+    abbr <- license_abbr("GPL", version, include_future)
+    use_description_field("License", abbr, overwrite = TRUE)
   }
-  use_license_template("GPL-3")
+  use_license_template(glue("GPL-{version}"))
 }
 
 #' @rdname licenses
 #' @export
-use_agpl3_license <- function() {
+use_agpl_license <- function(version = 3, include_future = TRUE) {
+  version <- check_license_version(version, 3)
+
   if (is_package()) {
-    use_description_field("License", "AGPL-3", overwrite = TRUE)
+    abbr <- license_abbr("AGPL", version, include_future)
+    use_description_field("License", abbr, overwrite = TRUE)
   }
-  use_license_template("AGPL-3")
+  use_license_template(glue("AGPL-{version}"))
 }
 
 #' @rdname licenses
 #' @export
-use_lgpl_license <- function() {
+use_lgpl_license <- function(version = 3, include_future = TRUE) {
+  version <- check_license_version(version, c(2.1, 3))
   if (is_package()) {
-    use_description_field("License", "LGPL (>= 2.1)", overwrite = TRUE)
+    abbr <- license_abbr("LGPL", version, include_future)
+    use_description_field("License", abbr, overwrite = TRUE)
   }
-  use_license_template("LGPL-2.1")
+  use_license_template(glue("LGPL-{version}"))
 }
 
 #' @rdname licenses
 #' @export
-use_apl2_license <- function() {
+use_apl_license <- function(version = 2, include_future = TRUE) {
+  version <- check_license_version(version, 2)
+
   if (is_package()) {
-    use_description_field("License", "Apache License (>= 2.0)", overwrite = TRUE)
+    abbr <- license_abbr("Apache License", version, include_future)
+    use_description_field("License", abbr, overwrite = TRUE)
   }
-  use_license_template("apache-2.0")
+  use_license_template(glue("apache-{version}"))
 }
 
 #' @rdname licenses
@@ -104,6 +131,32 @@ use_ccby_license <- function() {
   use_license_template("ccby-4")
 }
 
+
+# Fallbacks ---------------------------------------------------------------
+
+#' @rdname licenses
+#' @export
+#' @usage NULL
+use_gpl3_license <- function() {
+  use_gpl_license(3)
+}
+
+#' @rdname licenses
+#' @export
+#' @usage NULL
+use_agpl3_license <- function() {
+  use_agpl_license(3)
+}
+
+#' @rdname licenses
+#' @export
+#' @usage NULL
+use_apl2_license <- function() {
+  use_apl_license(2)
+}
+
+# Helpers -----------------------------------------------------------------
+
 use_license_template <- function(license, data = list()) {
   license_template <- glue("license-{license}.md")
 
@@ -112,4 +165,29 @@ use_license_template <- function(license, data = list()) {
     data = data,
     ignore = TRUE
   )
+}
+
+check_license_version <- function(version, possible) {
+  version <- as.double(version)
+
+  if (!version %in% possible) {
+    possible <- glue_collapse(possible, sep = "", last = "or, ")
+    abort(glue("`version` must be {possible}"))
+  }
+
+  version
+}
+
+license_abbr <- function(name, version, include_future) {
+  if (include_future) {
+    glue("{name} (>= {version})")
+  } else {
+    if (name %in% c("GPL", "LGPL", "AGPL")) {
+      # Standard abbreviations listed at
+      # https://cran.rstudio.com/doc/manuals/r-devel/R-exts.html#Licensing
+      glue("{name}-{version}")
+    } else {
+      glue("{name} (== {version})")
+    }
+  }
 }