+++ tests/testthat/test-license.R
@@ -14,10 +14,27 @@ test_that("use_mit_license() works", {
 test_that("other licenses work without error", {
   create_local_package()
 
-  expect_error(use_agpl3_license(), NA)
-  expect_error(use_apl2_license(), NA)
+  expect_error(use_agpl_license(3), NA)
+  expect_error(use_apl_license(2), NA)
   expect_error(use_cc0_license(), NA)
   expect_error(use_ccby_license(), NA)
+  expect_error(use_gpl_license(2), NA)
+  expect_error(use_gpl_license(3), NA)
+  expect_error(use_lgpl_license(2.1), NA)
+  expect_error(use_lgpl_license(3), NA)
+
+  # old fallbacks
+  expect_error(use_agpl3_license(), NA)
   expect_error(use_gpl3_license(), NA)
-  expect_error(use_lgpl_license(), NA)
+  expect_error(use_apl2_license(), NA)
+})
+
+test_that("check license gives useful errors", {
+  expect_error(check_license_version(1, 2), "must be 2")
+})
+
+test_that("generate correct abbreviations", {
+  expect_equal(license_abbr("GPL", 2, TRUE), "GPL (>= 2)")
+  expect_equal(license_abbr("GPL", 2, FALSE), "GPL-2")
+  expect_equal(license_abbr("Apache License", 2, FALSE), "Apache License (== 2)")
 })